Use c11 fixed size types

+++ Windsolar/file_reader.c
@@ -21,7 +21,7 @@ FileReader *FileReader_init(char *fname, size_t fsize)
     if (f == NULL)
         EXIT_WITH_MSG(EXIT_FAILURE, "Error: could not open %s\n", fname);
 
-    int16_t read_bytes = fread(fr->buff, sizeof(char), fsize, f);
+    size_t read_bytes = fread(fr->buff, sizeof(char), fsize, f);
     if (ferror(f) != 0 | read_bytes != fsize)
         EXIT_WITH_MSG(EXIT_FAILURE, "Error: could not read from %s\n", fname);
 
@@ -39,7 +39,7 @@ void FileReader_free(FileReader *fr)
     free(fr);
 }
 
-char FileReader_peek(FileReader *fr, int k)
+char FileReader_peek(FileReader *fr, int32_t k)
 {
     assert(fr);
     assert(k >= 0);
@@ -48,7 +48,7 @@ char FileReader_peek(FileReader *fr, int k)
     return fr->buff[fr->i + k];
 }
 
-char FileReader_consume(FileReader* fr, int k)
+char FileReader_consume(FileReader *fr, int32_t k)
 {
     char c = FileReader_peek(fr, k);
     fr->i += k + 1;
